Fix UseCaseConsultOrderContent

+++ Infrastructure/Ef/DbEntities/DbOrderHistoryContent.cs
@@ -2,7 +2,7 @@ namespace Infrastructure.Ef.DbEntities;
 
 public class DbOrderHistoryContent
 {
-    public double quantity { get; set; }
+    public decimal quantity { get; set; }
     
     public int idorder { get; set; }
